03/requirements.txt
mcp[cli]
openai
fastapi
python-dotenv
loguru # or use logging
requests
pytz
streamlit
websockets
websocket-client
pydantic
aiohttp
pyjwt
mysql-connector-python
aiomysql
